+++ purchases/src/main/kotlin/com/revenuecat/purchases/common/caching/DeviceCache.kt
@@ -134,6 +134,8 @@ public open class DeviceCache(
 
     private val offeringsResponseCacheKey: String by lazy { "$apiKeyPrefix.offeringsResponse" }
 
+    private val workflowsListResponseCacheKey: String by lazy { "$apiKeyPrefix.workflowsListResponse" }
+
     internal fun startEditing(): SharedPreferences.Editor {
         return preferences.edit()
     }
@@ -605,6 +607,27 @@ public open class DeviceCache(
     @Synchronized
     internal fun clearOfferingsResponseCache() {
         preferences.edit().remove(offeringsResponseCacheKey).apply()
+    }
+
+    // endregion
+
+    // region workflows list response
+
+    @Synchronized
+    internal fun getWorkflowsListResponseCache(): String? {
+        return preferences.getString(workflowsListResponseCacheKey, null)
+    }
+
+    @Synchronized
+    internal fun cacheWorkflowsListResponse(payload: String) {
+        preferences.edit()
+            .putString(workflowsListResponseCacheKey, payload)
+            .apply()
+    }
+
+    @Synchronized
+    internal fun clearWorkflowsListResponseCache() {
+        preferences.edit().remove(workflowsListResponseCacheKey).apply()
     }
 
     // endregion