+++ ui/revenuecatui/src/main/kotlin/com/revenuecat/purchases/ui/revenuecatui/customercenter/viewmodel/CustomerCenterViewModel.kt
@@ -1430,7 +1430,7 @@ internal class CustomerCenterViewModelImpl(
             val paywallArgs = PaywallActivityArgs(
                 offeringIdAndPresentedOfferingContext = OfferingSelection.IdAndPresentedOfferingContext(
                     offeringId = offering.identifier,
-                    presentedOfferingContext = offering.availablePackages.firstOrNull()?.presentedOfferingContext,
+                    presentedOfferingContext = offering.presentedOfferingContext,
                 ),
                 shouldDisplayDismissButton = true,
             )